fix(kt): use a Long counter in spread to avoid overflow at n>=65537
SpreadImpl's `active` pair counter was Int; for n >= 65537 the number of
pairs exceeds Int.MAX, wrapping it negative and tripping the `active <= 0`
guard, so a large constant vector returned -Infinity instead of 0 (the
other six languages return 0). The adjacent N and activeSize were already
Long. Add a regression test.

Acked-by: Andrey Akinshin <andrey.akinshin@gmail.com>

+++ kt/src/main/kotlin/dev/pragmastat/SpreadImpl.kt
@@ -88,13 +88,13 @@ internal fun spreadImpl(
         if (countBelow == prevCountBelow) {
             var minActive = Double.POSITIVE_INFINITY
             var maxActive = Double.NEGATIVE_INFINITY
-            var active = 0
+            var active = 0L
 
             for (i in 0 until n - 1) {
                 if (L[i] > R[i]) continue
                 minActive = minOf(minActive, a[L[i]] - a[i])
                 maxActive = maxOf(maxActive, a[R[i]] - a[i])
-                active += R[i] - L[i] + 1
+                active += (R[i] - L[i] + 1).toLong()
             }
 
             if (active <= 0) {
